api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sse-starlette>=1.8.0
websockets>=12.0
crewai>=0.28.0
crewai-tools>=0.1.0
google-generativeai>=0.3.0
requests>=2.31.0
APScheduler>=3.10.0
python-dateutil>=2.8.2

